requirements.txt
flask
psycopg2-binary
face-recognition
opencv-python
numpy
reportlab
werkzeug